Integrate logging package and refactor print statements
Replaced print statements with the logging package in home.dart, nutrisi.dart, suhu_kelembaban.dart, and background_service.dart for improved log management. Added the logging dependency to pubspec.yaml and updated imports and usage throughout the codebase. Also renamed Suhu_Kelembaban.dart to suhu_kelembaban.dart and updated references accordingly.
