+++ pubspec.yaml
@@ -63,6 +63,7 @@ dependencies:
   image_picker: ^1.1.1
   intl: ^0.19.0
   lottie: ^3.1.2
+  logging: ^1.2.0
   open_file: ^3.3.2
   path_provider: ^2.1.3
   percent_indicator: ^4.2.3
